% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat1.R
\name{cheat}
\alias{cheat}
\title{A function that helps you to cheat on the first assignment}
\usage{
cheat(x)
}
\arguments{
\item{x}{The question you want the answer for}
}
\value{
The answer to your question in quotation marks
}
\description{
A function that helps you to cheat on the first assignment
}
\examples{
cheat("Q1.1.7")
}
