% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind.R
\name{remind_me}
\alias{remind_me}
\title{A function that remind me of appointments}
\usage{
remind_me()
}
\value{
My appointments and their dates
}
\description{
A function that remind me of appointments
}
\examples{
remind_me()
}
